bugfix BG-1723-multisurface-in-createlegendinfo
This fix adds a switch to createLegendInfo to mask MultiSurface as Polygon.

The new ComplexTypes uses MultiSurface as geometry type. Before, MultiSurface
was added to the legend by its occurence. As "MultiSurface" did create an empty
set of subElements, the style v3 rule set was applied unordered.
Therefore ComplexTypes were shown as unsorted list in the legend. (This fix also
solves issue G31DEV1-164.)

With the new function getSubelementsFromXMLOtherStructure the empty set of
subElements got corrected and filled with MultiSurface geometries, which wasn't
processed by getSimpleGeometryStyle. So ComplexTypes with MultiSurface
geometries were left out of the legend.