BG-1718 bugfix filter layer in custom tree
The "featuresLoaded" event, triggert by VectorLayer, doesn't hand over the layer to
the createQuery method.

Using custom tree, the filter does not recognize a layer if it wasn't active before.
Starting the filter and then activate a layer for the first time, has no effect on
the filter. This is fixed now.