G31DEV1-220 Add: new attribute 'intersect' for sensor layer
For loading sensor data, current method is 'st_within', if the points
area is beyond the current map view, the points will disappear. But for
some sensor layer, the customer want to show the points if there are
intersects between the points area and map view.

A new attribute 'intersect' is implemented so that it could be
configured in config.json for sensor layer. The default value is ture
and if customer does not want to have a intersect, the value can be set
to false.