bugfix BG-1879-gfi-attributes-with-doublepoints
The sign ":" is a "magic" sign for i18next and can lead to misleading
values if we mix values and keys (as we do in default gfi theme).

This commit introduces the function "translateKeyWithPlausibilityCheck"
for the default gfi theme to avoid wrong interpretations of mixed
key/values, handed over to i18next translation method.

- if you pass a string with a : into i18next.t, the string is
  interpreted as key by i18next
  (e.g. $t("Test:") => "", as $t("common:unvalid.key") => "unvalid.key")

- if you pass a string with multiple : in it, i18next cuts the term
  before the first : and replaces all remaining : with .
  (e.g. $("Result: 2:1") => " 2.1")