BG-1874 Fix: setting the original title from meta data
It does not need to translate the title from meta data, normally the key
will be translated but the tile is not the key, just the information
from meta data.

If title is used as a translation, if it contains ":", the ":" wil be
recognized as nsSeparator to split namespace from key.

A new function translateKeyWithPlausibilityCheck is implemented as util
to prevent text with ":" to be recognized as translation key.