Restructure handling of objectPaths in styleModel
For adding a rotation parameter to a wfs feature it was neccessary to
use functions from the styleModel to get a property from an objectPath.
The getFeaturePropertyPath function is now an util function and includes
the isObjectPath function because they are being used exclusively together.
Their tests have also been moved out of the styleModel tests and are now
both within the unit tests for getFeaturePropertyPath.