g31DEV1-1165-bugfix-sliderRange-disco
Bugfix: Filter - SliderRange-Snippet

Using dataCustom as object to access with this.$options.dataCustom...,
the behavior of this variables are like class variables. They are
shared over components of the same type.
This caused interesting problems.

Now instance variables are declared in created() and bound to the
this-context, until we find a better solution.